Remove all references to "todo" and consolidate on "task"
Previously, the names "task" and "todo" were used in various places, sometimes
interchangeably and sometimes not. This consolidates on the name "task" (in
preparation for the "Task Planner") and removes all references to "todo".

The name "Todo" was *not* interchangeable in models, since there was both a
"Task" and a "Todo" struct. The "Task" struct was what actually existed in the
database, and the "Todo" struct was simply a holder for the form input. This was
renamed to "TaskInput" to better reflect its purpose.

In addition, all routes to "/todo" have been renamed to "/task"