Change task id to be required and non-nullable
After looking through the responses to diesel-rs/diesl #1440, I was able to get
this to work by having two different structs: Task and TaskWithId. This is
may change in the future, but for now looks like it will work.

The weirdest part was that I had to *not* include the #[table_name] macro for
the TaskWithId struct (I got a compiler error saying it couldn't find the macro
definition). However, when I removed it, it still worked, which was very
confusing.