Display plans along with tasks
This required changing the Context data structure that is passed to the
template. This also required changing the template to pull apart each successive
task and plans (the data was a vector of tuples of a task and its plans).

Also re-formatted the template with Emacs' jinja2-mode default indentation
style.